backend: rockchip: add libvpu backend
The libvpu is a backend access the Rockchip VPU V4L2 interface
with the help of libvpu. The main job of libvpu is analyze
the slice data and construct the frame information for it. It
is necessary in VDPAU day, so the VDPAU framework won't offer
enough information for decoder.

Signed-off-by: Randy Li <randy.li@rock-chips.com>