backend: v4l2: getting those VA-API not offered settings from librkdec
We found there is still some extra data required by the HW to decode
H.264 video.

Reading this commit would know what the VA-API could and couldn't offer
in currently version.

Notice that, the flags in sps would apply to hardware as well, but I can't
find anything different in my test.

With the "media: rockchip-vpu: fix the wrong size of slice_param for H.264"
The slice_param would be sent fix size to kernel.

The performance is not 100 present as the v4l2_libvpu, there would be some
noise shake in playing and I don't know why.

Signed-off-by: Randy Li <randy.li@rock-chips.com>